Refactory Modal
