--- src/components/folderPageComponents/modal/Modal.js
@@ -1,5 +1,0 @@
-import { ModalForm } from './ModalForm';
-import React from 'react';
-export const Modal = ({ modalData }) => {
-  return <ModalForm modalData={modalData} />;
-};
