+++ src/components/modal/DeleteFolderModal.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import { RedButton } from '../../styles/styledComponents/folderStyled';
+
+export const DeleteFolderModal = () => {
+  return (
+    <>
+      <RedButton>삭제하기</RedButton>
+    </>
+  );
+};
