+++ src/components/modal/DeleteLinkModal.js
@@ -1,0 +1,9 @@
+import React from 'react';
+import { RedButton } from '../../styles/styledComponents/folderStyled';
+export const DeleteLinkModal = () => {
+  return (
+    <>
+      <RedButton>삭제하기</RedButton>
+    </>
+  );
+};
