+++ src/components/modal/EditFolderModal.js
@@ -1,0 +1,13 @@
+import React from 'react';
+import {
+  ModalInput,
+  BlueButton,
+} from '../../styles/styledComponents/folderStyled';
+export const EditFolderModal = () => {
+  return (
+    <>
+      <ModalInput type="text" placeholder="유용한 팁" />
+      <BlueButton>변경하기</BlueButton>
+    </>
+  );
+};
